adding images and updating files
